% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{SeuratBasics-package}
\alias{SeuratBasics-package}
\title{Tools for single-cell genomics}
\description{
This is a slimmed down version of Seurat, only keeping functionalities deemed
the most commonly used.
}
\section{Package options}{


Seurat uses the following [options()] to configure behaviour:

\describe{
  \item{\code{Seurat.memsafe}}{global option to call gc() after many operations.
  This can be helpful in cleaning up the memory status of the R session and
  prevent use of swap space. However, it does add to the computational overhead
  and setting to FALSE can speed things up if you're working in an environment
  where RAM availability is not a concern.}
  \item{\code{Seurat.checkdots}}{For functions that have ... as a parameter,
  this controls the behavior when an item isn't used. Can be one of warn,
  stop, or silent.}
}
}

